# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/merchant/store": {
      "get": {
        "operationId": "MerchantController_store",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "patch": {
        "operationId": "MerchantController_updateStore",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/dashboard": {
      "get": {
        "operationId": "MerchantController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/categories": {
      "get": {
        "operationId": "MerchantController_categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "post": {
        "operationId": "MerchantController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/categories/{id}": {
      "put": {
        "operationId": "MerchantController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "delete": {
        "operationId": "MerchantController_deleteCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/products": {
      "get": {
        "operationId": "MerchantController_products",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "post": {
        "operationId": "MerchantController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/products/{id}": {
      "patch": {
        "operationId": "MerchantController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "delete": {
        "operationId": "MerchantController_deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/products/import-from-catalog": {
      "post": {
        "operationId": "MerchantController_importFromCatalog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportFromCatalogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/offers": {
      "get": {
        "operationId": "MerchantController_offers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "post": {
        "operationId": "MerchantController_createOffer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/offers/{id}": {
      "patch": {
        "operationId": "MerchantController_updateOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferPatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "delete": {
        "operationId": "MerchantController_deleteOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/pending-customers": {
      "get": {
        "operationId": "MerchantController_pendingCustomers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      },
      "post": {
        "operationId": "MerchantController_createPendingCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PendingCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/custom-requests": {
      "get": {
        "operationId": "MerchantController_customRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/custom-requests/{id}/quote": {
      "post": {
        "operationId": "MerchantController_quote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/custom-requests/{id}/reject": {
      "post": {
        "operationId": "MerchantController_rejectRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/reports/summary": {
      "get": {
        "operationId": "MerchantController_reportSummary",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/reports/orders": {
      "get": {
        "operationId": "MerchantController_reportOrders",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/pos/snapshot": {
      "get": {
        "operationId": "MerchantController_snapshot",
        "parameters": [
          {
            "name": "since",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/pos/sales": {
      "post": {
        "operationId": "MerchantController_posSale",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/merchant/pos/products": {
      "post": {
        "operationId": "MerchantController_posProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "merchant"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wasl API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateStoreDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "area": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "delivery_info": {
            "type": "string",
            "maxLength": 300,
            "nullable": true
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "nullable": true
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "maxLength": 500,
            "nullable": true
          },
          "is_open": {
            "type": "boolean"
          },
          "lat": {
            "type": "number",
            "minimum": -90,
            "exclusiveMinimum": false,
            "maximum": 90,
            "exclusiveMaximum": false,
            "nullable": true
          },
          "lng": {
            "type": "number",
            "minimum": -180,
            "exclusiveMinimum": false,
            "maximum": 180,
            "exclusiveMaximum": false,
            "nullable": true
          },
          "business_category_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "sort_order": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 10000000,
            "exclusiveMaximum": false
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "barcode": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "in_stock": {
            "type": "boolean"
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "lib_category": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "main_section": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "subcategory": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          }
        },
        "required": [
          "name",
          "price",
          "image_url"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 10000000,
            "exclusiveMaximum": false
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "barcode": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "in_stock": {
            "type": "boolean"
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "lib_category": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "main_section": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "subcategory": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          }
        }
      },
      "ImportFromCatalogDto": {
        "type": "object",
        "properties": {
          "catalog_item_ids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 500,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "catalog_item_ids"
        ]
      },
      "OfferDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "discount_price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "max_qty": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "product_id",
          "discount_price"
        ]
      },
      "OfferPatchDto": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "discount_price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "max_qty": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "PendingCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "phone": {
            "type": "string",
            "minLength": 6,
            "maxLength": 20
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
      "QuoteDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "note": {
            "type": "string",
            "maxLength": 300,
            "nullable": true
          }
        },
        "required": [
          "price"
        ]
      },
      "PosSaleDto": {
        "type": "object",
        "properties": {
          "op_id": {
            "type": "string",
            "format": "uuid"
          },
          "order": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "customer_id": {
                "type": "string",
                "format": "uuid",
                "nullable": true
              },
              "payment_method": {
                "type": "string",
                "enum": [
                  "cash",
                  "credit",
                  "wallet",
                  "jeeb",
                  "jawali",
                  "hasab",
                  "onecash"
                ]
              },
              "note": {
                "type": "string",
                "maxLength": 500,
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "id",
              "payment_method",
              "created_at"
            ]
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "maxItems": 200,
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "product_id": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "name": {
                  "type": "string",
                  "maxLength": 160
                },
                "price": {
                  "type": "number",
                  "minimum": 0,
                  "exclusiveMinimum": false
                },
                "qty": {
                  "type": "integer",
                  "minimum": 1,
                  "exclusiveMinimum": false
                },
                "note": {
                  "type": "string",
                  "maxLength": 200,
                  "nullable": true
                }
              },
              "required": [
                "product_id",
                "name",
                "price",
                "qty"
              ]
            }
          },
          "credit": {
            "type": "object",
            "properties": {
              "customerKind": {
                "type": "string",
                "enum": [
                  "registered",
                  "pending"
                ]
              },
              "customerId": {
                "type": "string",
                "format": "uuid"
              },
              "txId": {
                "type": "string",
                "format": "uuid"
              },
              "amount": {
                "type": "number",
                "minimum": 0,
                "exclusiveMinimum": true
              },
              "note": {
                "type": "string",
                "maxLength": 300,
                "nullable": true
              }
            },
            "required": [
              "customerKind",
              "customerId",
              "txId",
              "amount"
            ],
            "nullable": true
          }
        },
        "required": [
          "op_id",
          "order",
          "items"
        ]
      },
      "PosProductDto": {
        "type": "object",
        "properties": {
          "op_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 10000000,
            "exclusiveMaximum": false
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "barcode": {
            "type": "string",
            "maxLength": 64,
            "nullable": true
          },
          "in_stock": {
            "type": "boolean"
          },
          "image_url": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "lib_category": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "main_section": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "subcategory": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "image_url"
        ]
      },
...
    }
  }
}